% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap_r}
\alias{bootstrap_loop}
\alias{run_benchmarks}
\title{Bootstraps}
\usage{
bootstrap_r(x, R)

bootstrap_loop(x, R)

run_benchmarks(x = runif(2000), R = 2000, times = 10, ...)
}
\arguments{
\item{x}{A vector of numbers}

\item{R}{Number of replications}

\item{times}{Number of times to run benchmarks}

\item{...}{Additional arguments passed to \code{\link[microbenchmark:microbenchmark]{microbenchmark::microbenchmark()}}}
}
\description{
Bootstrap functions in R, C++, and rust
}
\examples{
bootstrap_r(runif(100), 100)
}
