% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\alias{bootstrap_r}
\alias{bootstrap_loop}
\alias{bootstrap_rs}
\alias{run_benchmarks}
\alias{main}
\title{Bootstraps}
\usage{
bootstrap_r(x, R)

bootstrap_loop(x, R)

bootstrap_rs(x, R)

run_benchmarks(x = runif(2000), R = 2000, times = 10, ...)

main()
}
\arguments{
\item{x}{A vector of numbers}

\item{R}{Number of replications}

\item{times}{Number of times to run benchmarks}

\item{...}{Additional arguments passed to [microbenchmark::microbenchmark()]}
}
\description{
Bootstrap functions in R and Rust
}
\examples{
\dontrun{
numbers <- runif(100)
resamples <- 100
bootstrap_r(numbers, resamples)
bootstrap_loop(numbers, resamples)
bootstrap_rs(numbers, resamples)
}
}
